+++ mrb_doc/models/image.rb
@@ -133,6 +133,12 @@ class Image
     self
   end
 
+  # Generates mipmaps for the image
+  # @return [Image]
+  def generate_mipmaps!
+    # src/mruby_integration/models/image.cpp
+    self
+  end
 
   # Generates a new image of width by height in the specified colour.
   # @param width [Integer]