+++ src/mruby_integration/models/image.cpp
@@ -150,6 +150,11 @@ void append_models_Image(mrb_state *mrb) {
         self
       end
 
+      def generate_mipmaps!
+        image_mipmaps!(self)
+        self
+      end
+
       def self.generate(width:, height:, colour: RAYWHITE)
         generate_image_colour(width, height, colour)
       end